MDEV-30889: max_global_memory_used status variable
Implement a new status variable max_global_memory_used, which gives the peak
value of global_memory_used.

When mallinfo2() shows a lot of memory being free in the malloc
implementation, this can be used to understand if this is because that
memory was at one point allocated to the server during a peak load situation
and later freed; or if it is caused by internal fragmentation in the malloc
implementation.

Signed-off-by: Kristian Nielsen <knielsen@knielsen-hq.org>